data-science/EDA-Notebooks/requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
psycopg2-binary==2.9.7
plotly==6.2.0
sqlalchemy==2.0.41
streamlit==1.47.0